docker/apps/python-app/pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary==2.9.10",
]